CRC64 JS implementation

+++ sdk/storage/storage-blob/test/blobclientpollers.spec.ts
@@ -38,7 +38,7 @@ describe("BlobClient beginCopyFromURL Poller", () => {
   });
 
   afterEach(async function (this: Context) {
-    if (!this.currentTest?.isPending()) {
+    if (!(this.currentTest && this.currentTest.isPending())) {
       await containerClient.delete();
       await destinationContainerClient.delete();
       await recorder.stop();
