+++ sdk/storage/storage-blob/test/node/blockblobclient.spec.ts
@@ -53,7 +53,7 @@ describe("BlockBlobClient Node.js only", () => {
   });
 
   afterEach(async function (this: Context) {
-    if (!this.currentTest?.isPending()) {
+    if (!(this.currentTest && this.currentTest.isPending())) {
       await containerClient.delete();
       await recorder.stop();
     }
@@ -238,7 +238,7 @@ describe("syncUploadFromURL", () => {
   });
 
   afterEach(async function (this: Context) {
-    if (!this.currentTest?.isPending()) {
+    if (!(this.currentTest && this.currentTest.isPending())) {
       await containerClient.delete();
       await recorder.stop();
     }
