+++ sdk/storage/storage-blob/test/node/emulator-tests.spec.ts
@@ -39,7 +39,7 @@ describe("Emulator Tests", () => {
   });
 
   afterEach(async function (this: Context) {
-    if (!this.currentTest?.isPending()) {
+    if (!(this.currentTest && this.currentTest.isPending())) {
       await containerClient.delete();
     }
   });
