+++ sdk/storage/storage-blob/test/node/highlevel.node.spec.ts
@@ -57,7 +57,7 @@ describe("Highlevel", () => {
   });
 
   afterEach(async function (this: Context) {
-    if (!this.currentTest?.isPending()) {
+    if (!(this.currentTest && this.currentTest.isPending())) {
       await containerClient.delete();
       await recorder.stop();
     }
